requirements.txt
Django==1.4.2
MySQL-python==1.2.5
South==0.7.5
django-admin-jqueryui==1.8.22
django-allauth==0.4.0
django-email-confirmation==0.2
django-extensions==0.9
django-google-analytics==1
django-keyedcache==1.4-1
django-livesettings==1.4-7
django-modeltranslation==0.3.3
django-registration==0.7
django-rosetta==0.6.2
django-timezone-field==1.0
django-tinymce==1.5.1b2
httplib2==0.7.4
ipython==1.2.1
lxml==2.2.4
oauth2==1.5.211
pycrypto==2.4
pycurl==7.19.0
python-dateutil==2.2
python-openid==2.2.5
pytz==2014.4
six==1.5.2
wsgiref==0.1.2

